checkpoint: reworked posted on, last refreshed date

+++ api/main.py
@@ -6,7 +6,7 @@ from typing import List
 from fastapi import FastAPI, HTTPException
 from fastapi.middleware.cors import CORSMiddleware
 from sqlalchemy import func
-from sqlmodel import Session, col, desc, or_, select
+from sqlmodel import Session, case, col, desc, or_, select
 
 from database import get_engine
 from models import (
@@ -48,14 +48,21 @@ def get_sources():
         if not categories:
             raise HTTPException(status_code=404, detail="No categories found")
 
+        current_time = datetime.now()
         for category in categories:
+            last_refreshed_string = "never"
+            if category.last_refreshed:
+                last_refreshed_delta = current_time - category.last_refreshed
+                hours_ago = int(last_refreshed_delta.total_seconds() / 3600)
+                last_refreshed_string = f"{hours_ago} hours ago"
+
             data = {
                 # ignored err, company is guaranteed to exist by design, sqlmodel quirk forces optional type in definition
                 "company_name": category.company.name,  # type: ignore
                 "category_id": category.id,
                 "category_name": category.name,
                 "url": category.url,
-                "last_refreshed": category.last_refreshed,
+                "last_refreshed": last_refreshed_string,
             }
             result.append(data)
 
@@ -63,18 +70,25 @@ def get_sources():
 
 
 @app.post("/submit_jobs", response_model=SubmitJobsResponseDTO)
-def submit_jobs(jobs: list[ScrapedJobDTO]):
+def submit_jobs(jobs: list[ScrapedJobDTO], first_batch: bool = False):
     with Session(engine) as session:
         submit_jobs_response = SubmitJobsResponseDTO(created_count=0)
-        session.delete(Job)
+
+        if first_batch:
+            session.delete(Job)
+            session.commit()
 
         updated_categories = set()
 
         for job in jobs:
+            cleaned_date = None
+            if job.date and job.date != "None":
+                cleaned_date = datetime.fromisoformat(job.date)
+
             new_job = Job(
                 title=job.title,
                 category_id=job.category_id,
-                date=job.date,
+                date=cleaned_date,
             )
             session.add(new_job)
             submit_jobs_response.created_count += 1
@@ -95,6 +109,7 @@ def submit_jobs(jobs: list[ScrapedJobDTO]):
 def search_jobs(search_params: JobSearchParamsDTO):
     with Session(engine) as session:
         results = []
+
         statement = select(Job).order_by(desc(Job.date))
 
         if search_params.keywords:
@@ -120,14 +135,18 @@ def search_jobs(search_params: JobSearchParamsDTO):
         if not jobs:
             raise HTTPException(status_code=404, detail="No jobs found")
 
+        current_time = datetime.now()
         for job in jobs:
+            last_refreshed_delta = current_time - job.category.last_refreshed  # type: ignore
+            hours_ago = int(last_refreshed_delta.total_seconds() / 3600)
+
             jobSearchResult = JobSearchResultDTO(
                 title=job.title,
                 company=job.category.company.name,  # type: ignore
                 category=job.category.name,  # type: ignore
                 url=job.category.url,  # type: ignore
                 date=job.date,
-                last_refreshed=job.category.last_refreshed,  # type: ignore
+                last_refreshed=f"{hours_ago} hours ago",
             )
             results.append(jobSearchResult)
